Adicionado fluxo de notificação, gerar notificação ao enviar pedido, expirar pedido. Adicionado opção de entrar em fila de espera para produto esgotado, enviar notificação ao voltar ao estoque. Adicionado algumas migrations de insert de produtos e usuarios padrao

+++ EcommerceV3/src/main/java/com/ec/ecommercev3/Repository/ProductRepository.java
@@ -21,7 +21,7 @@ public interface ProductRepository extends JpaRepository<Product, Long>,
             "UPDATE _product SET stock = stock + :stockQuantityDiscount WHERE id = :productId")
     void updateStock(
             @Param("productId") Long productId,
-            @Param("stockQuantityDiscount") Integer stockQuantityDiscount);
+            @Param("stockQuantityDiscount") Long stockQuantityDiscount);
